requirements.txt
fastapi
uvicorn
pyyaml
ansible-runner
requests
